BAU: move out lambda results helper to repo level for all stubs
